[PHP 7] Update tests for Bolt\Helpers\String to Bolt\Helpers\Str rename
